Move Wrap to errutil package
