+++ pkg/services/dashboards/dashboard_service.go
@@ -9,6 +9,7 @@ import (
 	"github.com/grafana/grafana/pkg/models"
 	"github.com/grafana/grafana/pkg/services/guardian"
 	"github.com/grafana/grafana/pkg/util"
+	"github.com/grafana/grafana/pkg/util/errutil"
 )
 
 // DashboardService service for operating on dashboards
@@ -260,7 +261,7 @@ func (dr *dashboardServiceImpl) deleteDashboard(dashboardId int64, orgId int64, 
 		isDashboardProvisioned := &models.IsDashboardProvisionedQuery{DashboardId: dashboardId}
 		err := bus.Dispatch(isDashboardProvisioned)
 		if err != nil {
-			return util.Wrap("failed to check if dashboard is provisioned", err)
+			return errutil.Wrap("failed to check if dashboard is provisioned", err)
 		}
 
 		if isDashboardProvisioned.Result {
